Refactor code to use modules
