+++ src/modules/task.js
@@ -1,0 +1,8 @@
+//Create a task class
+export default class Task {
+    constructor(description, completed, index) {
+        this.description = description;
+        this.completed = completed;
+        this.index = index;
+    }
+}
